+++ src/ggc.c
@@ -2,11 +2,6 @@
 #include "interpreter.h"
 #include "type.h"
 
-#define REACHABLE(ob) ((ob)->flags & Nst_FLAG_GCC_REACHABLE)
-#define UNREACHABLE(ob) ((ob)->flags & Nst_FLAG_GCC_UNREACHABLE)
-
-static bool is_init = false;
-
 static inline void move_obj(Nst_GGCObj *obj, Nst_GGCList *from, Nst_GGCList *to)
 {
     if (from->len == 1) {
@@ -218,7 +213,7 @@ void _Nst_ggc_delete_objs(void)
     free_obj_memory(&Nst_state.ggc.old_gen);
 }
 
-void Nst_ggc_init(void)
+void _Nst_ggc_init(void)
 {
     Nst_GGCList gen1 = { NULL, NULL, 0 };
     Nst_GGCList gen2 = { NULL, NULL, 0 };
@@ -229,13 +224,6 @@ void Nst_ggc_init(void)
     Nst_state.ggc.gen3 = gen3;
     Nst_state.ggc.old_gen = old_gen;
     Nst_state.ggc.old_gen_pending = 0;
-
-    is_init = true;
-}
-
-bool Nst_ggc_was_init(void)
-{
-    return is_init;
 }
 
 void _Nst_ggc_obj_reachable(Nst_Obj *obj)